feat(impl):添加了user的异常报错

+++ src/main/java/local/happysixplus/backendcodeanalysis/service/impl/AdminUserServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import local.happysixplus.backendcodeanalysis.data.AdminUserData;
+import local.happysixplus.backendcodeanalysis.exception.MyRuntimeException;
 import local.happysixplus.backendcodeanalysis.po.AdminUserPo;
 import local.happysixplus.backendcodeanalysis.service.AdminUserService;
 import local.happysixplus.backendcodeanalysis.vo.AdminUserVo;
@@ -20,16 +21,22 @@ public class AdminUserServiceImpl implements AdminUserService {
 
     @Override
     public void addAdmin(AdminUserVo vo) {
+        if (adminUserData.findByUsername(vo.getUsername()) != null)
+            throw new MyRuntimeException("该用户名已被注册");
         adminUserData.save(new AdminUserPo(null, vo.getUsername(), vo.getPwdMd5()));
     };
 
     @Override
     public void removeAdmin(Long id) {
-        adminUserData.deleteById(id);
+        if (adminUserData.existsById(id))
+            adminUserData.deleteById(id);
+        else
+            throw new MyRuntimeException("该管理员不存在");
     };
 
     @Override
     public void updateAdmin(AdminUserVo vo) {
+
         adminUserData.save(new AdminUserPo(vo.getId(), vo.getUsername(), vo.getPwdMd5()));
     };
 
@@ -45,6 +52,8 @@ public class AdminUserServiceImpl implements AdminUserService {
     @Override
     public AdminUserVo getOneAdmin(Long id) {
         var po = adminUserData.findById(id).orElse(null);
+        if (po == null)
+            throw new MyRuntimeException("该管理员不存在");
         return new AdminUserVo(po.getId(), po.getUsername(), po.getPwdMd5(), null);
     };
 }
